+++ packages/kitchen/src/types/theme.ts
@@ -1,5 +1,5 @@
 export type Theme = {
-  id: number;
+  name: string;
   colors: Colors;
 };